# bff/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Basalt Arena BFF API
  version: 1.0.0
  description: REST API used by the public arena client and the custom admin panel.
servers:
  - url: http://localhost:3001
...
paths:
...
  /api/v1/sprints:
    get:
      tags: [sprints]
      summary: List published, non-archived sprints with solutions
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SortBy'
      responses:
        '200':
          description: List of sprints
  /api/v1/sprints/{id}:
    get:
      tags: [sprints]
      summary: Sprint detail with solutions
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/SortBy'
      responses:
        '200':
          description: Sprint detail
        '404':
          description: Sprint not found or not public
  /api/v1/sprints/{id}/solutions:
    get:
      tags: [sprints]
      summary: Sprint solutions with sorting
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/SortBy'
      responses:
        '200':
          description: Solutions list
  /api/v1/sprints/{id}/submissions:
    post:
      tags: [sprints]
      summary: Submit or update current user's solution for a sprint
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionUpsertRequest'
      responses:
        '201':
          description: Created or updated submission
...
  /api/v1/admin/users/{userId}/sprints:
    get:
      tags: [admin]
      summary: List sprint access and submission activity for a user
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User sprint activity
  /api/v1/admin/sprints:
    get:
      tags: [admin]
      summary: List all sprints, including drafts and archived records
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Sprints
    post:
      tags: [admin]
      summary: Create sprint
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSprintWriteRequest'
      responses:
        '201':
          description: Created sprint
  /api/v1/admin/sprints/{id}:
    patch:
      tags: [admin]
      summary: Update sprint fields
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSprintPatchRequest'
      responses:
        '200':
          description: Updated sprint
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
    UserId:
      in: path
      name: userId
      required: true
      schema: { type: string }
...
    SortBy:
      in: query
      name: sortBy
      schema:
        type: string
        enum: [efficiency, likes, mentor]
        default: efficiency
...
  schemas:
...
    SubmissionUpsertRequest:
      type: object
      required: [repoUrl]
      additionalProperties: false
      properties:
        repoUrl: { type: string, format: uri, maxLength: 2048 }
        demoUrl:
          oneOf:
            - { type: string, format: uri, maxLength: 2048 }
            - { type: string, const: '' }
...
    AdminSprintWriteRequest:
      type: object
      required: [slug, title, tabLabel, completedLabel]
      additionalProperties: false
      properties:
        slug: { type: string, pattern: '^[a-z0-9-]+$', maxLength: 128 }
        title: { type: string, minLength: 1, maxLength: 256 }
        tabLabel: { type: string, minLength: 1, maxLength: 128 }
        tabIcon:
          oneOf:
            - { type: string, maxLength: 64 }
            - { type: 'null' }
        completedLabel: { type: string, minLength: 1, maxLength: 128 }
        tags:
          type: array
          maxItems: 50
          items: { type: string, maxLength: 64 }
        active: { type: boolean }
        published: { type: boolean }
        archived: { type: boolean }
        startsAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        endsAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        brief:
          type: object
          additionalProperties: true
        metrics:
          type: object
          additionalProperties: true
    AdminSprintPatchRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        slug: { type: string, pattern: '^[a-z0-9-]+$', maxLength: 128 }
        title: { type: string, minLength: 1, maxLength: 256 }
        tabLabel: { type: string, minLength: 1, maxLength: 128 }
        tabIcon:
          oneOf:
            - { type: string, maxLength: 64 }
            - { type: 'null' }
        completedLabel: { type: string, minLength: 1, maxLength: 128 }
        tags:
          type: array
          maxItems: 50
          items: { type: string, maxLength: 64 }
        active: { type: boolean }
        published: { type: boolean }
        archived: { type: boolean }
        startsAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        endsAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        brief:
          type: object
          additionalProperties: true
        metrics:
          type: object
          additionalProperties: true
...